More of Chapter 14
I tried to work in this chapter more but for some reason my code keeps giving me an error "collect2.exe: error: ld returned 1 exit status" and that it can't find the reference. I've got the exact same code as the book, the book code won't even work. I don't know if it is a vs code issue or if I have the wrong compiler but this is making me frustrated. For the rest of the chapter I'll create classes in the main program that is running, hopefully I can figure out why this is happening.